refactor: move footer and matomowrapper to components
