+++ apps/events-helsinki/src/constants.ts
@@ -1,5 +1,3 @@
-import type { Language } from './types';
-
 export const DEFAULT_LANGUAGE = 'fi';
 
 export const ROUTES = {
@@ -52,28 +50,4 @@ export const FEEDBACK_LINKS = {
   fi: 'https://www.hel.fi/helsinki/fi/kaupunki-ja-hallinto/osallistu-ja-vaikuta/palaute',
   en: 'https://www.hel.fi/helsinki/en/administration/participate/feedback',
   sv: 'https://www.hel.fi/helsinki/sv/stad-och-forvaltning/delta/feedback',
-};
-
-export const DEFAULT_HEADER_MENU_NAME: Record<Language, string> = {
-  fi:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_FI ??
-    'Events Helsinki Header FI',
-  en:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_EN ??
-    'Events Helsinki Header EN',
-  sv:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_SV ??
-    'Events Helsinki Header SV',
-};
-
-export const DEFAULT_FOOTER_MENU_NAME: Record<Language, string> = {
-  fi:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_FI ??
-    'Events Helsinki Footer FI',
-  en:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_EN ??
-    'Events Helsinki Footer EN',
-  sv:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_SV ??
-    'Events Helsinki Footer SV',
 };
