--- apps/events-helsinki/src/domain/footer/footerCategories.module.scss
@@ -1,45 +1,0 @@
-@import 'src/styles/breakpoints';
-@import 'src/styles/mixins';
-
-.divider {
-  background-color: var(--footer-divider-color);
-  border: 0;
-  height: 1px;
-  margin: 0 0 var(--spacing-xs);
-  width: 100%;
-}
-
-.topFooterWrapper {
-  margin: 0 var(--spacing-m);
-  padding: 0 0 3rem;
-  display: grid;
-  align-items: center;
-}
-
-.categoriesTitle {
-  font-size: 2rem;
-  font-weight: 600;
-  margin-bottom: 1rem;
-}
-
-.categoriesInnerWrapper {
-  padding: 1.5rem 0 0;
-  display: grid;
-  align-items: center;
-  height: 100%;
-  grid-template-columns: 1fr;
-  grid-gap: 1rem;
-
-  @include respond-above(s) {
-    grid-template-columns: repeat(2, 1fr);
-    grid-gap: 1rem;
-  }
-
-  @include respond-above(m) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  @include respond-above(xl) {
-    grid-template-columns: repeat(5, 1fr);
-  }
-}
