--- apps/hobbies-helsinki/src/domain/matomoWrapper/MatomoWrapper.tsx
@@ -1,24 +1,0 @@
-import { useMatomo } from '@jonkoops/matomo-tracker-react';
-import { useRouter } from 'next/router';
-import type { ReactNode } from 'react';
-import React, { useEffect } from 'react';
-
-interface Props {
-  children?: ReactNode | undefined;
-}
-
-const MatomoWrapper: React.FC<Props> = ({ children }) => {
-  const { trackPageView } = useMatomo();
-  const { asPath: pathname } = useRouter();
-
-  // Track page changes when pathnname changes
-  useEffect(() => {
-    trackPageView({
-      href: window.location.href,
-    });
-  }, [pathname, trackPageView]);
-
-  return <>{children}</>;
-};
-
-export default MatomoWrapper;
