+++ apps/hobbies-helsinki/src/pages/index.tsx
@@ -3,7 +3,10 @@ import {
   getQlLanguage,
   NavigationContext,
   Navigation,
+  MatomoWrapper,
+  useCommonTranslation,
 } from 'events-helsinki-components';
+import FooterSection from 'events-helsinki-components/components/footer/Footer';
 import type { GetStaticPropsContext, NextPage } from 'next';
 import React, { useContext } from 'react';
 import type { PageType, ArticleType } from 'react-helsinki-headless-cms';
@@ -26,9 +29,7 @@ import {
 import getHobbiesStaticProps from '../domain/app/getHobbiesStaticProps';
 import cmsHelper from '../domain/app/headlessCmsHelper';
 import routerHelper from '../domain/app/routerHelper';
-import FooterSection from '../domain/footer/Footer';
 import serverSideTranslationsWithCommon from '../domain/i18n/serverSideTranslationsWithCommon';
-import MatomoWrapper from '../domain/matomoWrapper/MatomoWrapper';
 import { LandingPageContentLayout } from '../domain/search/landingPage/LandingPage';
 
 const HomePage: NextPage<{
@@ -40,7 +41,7 @@ const HomePage: NextPage<{
     utils: { getRoutedInternalHref },
   } = useConfig();
   const { footerMenu } = useContext(NavigationContext);
-
+  const { t } = useCommonTranslation();
   return (
     <MatomoWrapper>
       <HCRCPage
@@ -57,7 +58,9 @@ const HomePage: NextPage<{
             language={getQlLanguage(locale)}
           />
         }
-        footer={<FooterSection menu={footerMenu} />}
+        footer={
+          <FooterSection menu={footerMenu} appName={t('appHobbies:appName')} />
+        }
       />
     </MatomoWrapper>
   );
