+++ apps/hobbies-helsinki/src/pages/pages/[...slug].tsx
@@ -2,7 +2,13 @@
 
 import type { NormalizedCacheObject } from '@apollo/client';
 import type { AppLanguage } from 'events-helsinki-components';
-import { NavigationContext, getAllPages } from 'events-helsinki-components';
+import {
+  NavigationContext,
+  getAllPages,
+  MatomoWrapper,
+  useCommonTranslation,
+} from 'events-helsinki-components';
+import FooterSection from 'events-helsinki-components/components/footer/Footer';
 import type {
   GetStaticPropsContext,
   GetStaticPropsResult,
@@ -32,9 +38,7 @@ import getHobbiesStaticProps from '../../domain/app/getHobbiesStaticProps';
 import cmsHelper from '../../domain/app/headlessCmsHelper';
 import routerHelper from '../../domain/app/routerHelper';
 import { apolloClient } from '../../domain/clients/eventsFederationApolloClient';
-import FooterSection from '../../domain/footer/Footer';
 import serverSideTranslationsWithCommon from '../../domain/i18n/serverSideTranslationsWithCommon';
-import MatomoWrapper from '../../domain/matomoWrapper/MatomoWrapper';
 
 const NextCmsPage: NextPage<{
   page: PageType;
@@ -45,7 +49,7 @@ const NextCmsPage: NextPage<{
     utils: { getRoutedInternalHref },
   } = useConfig();
   const { footerMenu } = useContext(NavigationContext);
-
+  const { t } = useCommonTranslation();
   return (
     <MatomoWrapper>
       <HCRCPage
@@ -61,7 +65,9 @@ const NextCmsPage: NextPage<{
             }
           />
         }
-        footer={<FooterSection menu={footerMenu} />}
+        footer={
+          <FooterSection menu={footerMenu} appName={t('appHobbies:appName')} />
+        }
       />
     </MatomoWrapper>
   );
