+++ apps/hobbies-helsinki/src/pages/search/index.tsx
@@ -2,7 +2,9 @@ import {
   NavigationContext,
   useAppHobbiesTranslation,
   Navigation,
+  MatomoWrapper,
 } from 'events-helsinki-components';
+import FooterSection from 'events-helsinki-components/components/footer/Footer';
 import type { GetStaticPropsContext } from 'next';
 import { useRouter } from 'next/router';
 import React, { useRef, useEffect, useContext } from 'react';
@@ -10,9 +12,7 @@ import { Page as HCRCApolloPage } from 'react-helsinki-headless-cms/apollo';
 import { ROUTES } from '../../constants';
 import getHobbiesStaticProps from '../../domain/app/getHobbiesStaticProps';
 import routerHelper from '../../domain/app/routerHelper';
-import FooterSection from '../../domain/footer/Footer';
 import serverSideTranslationsWithCommon from '../../domain/i18n/serverSideTranslationsWithCommon';
-import MatomoWrapper from '../../domain/matomoWrapper/MatomoWrapper';
 import AdvancedSearch from '../../domain/search/eventSearch/AdvancedSearch';
 import SearchPage from '../../domain/search/eventSearch/SearchPage';
 
@@ -51,7 +51,12 @@ export default function Search() {
             pageTitle={tAppHobbies('appHobbies:search.pageTitle')}
           />
         }
-        footer={<FooterSection menu={footerMenu} />}
+        footer={
+          <FooterSection
+            menu={footerMenu}
+            appName={tAppHobbies('appHobbies:appName')}
+          />
+        }
       />
     </MatomoWrapper>
   );
