+++ apps/sports-helsinki/src/constants.ts
@@ -1,5 +1,3 @@
-import type { AppLanguage } from 'events-helsinki-components';
-
 export const DEFAULT_LANGUAGE = 'fi';
 
 export const SEARCH_ROUTES = {
@@ -60,28 +58,4 @@ export const FEEDBACK_LINKS = {
   fi: 'https://www.hel.fi/helsinki/fi/kaupunki-ja-hallinto/osallistu-ja-vaikuta/palaute',
   en: 'https://www.hel.fi/helsinki/en/administration/participate/feedback',
   sv: 'https://www.hel.fi/helsinki/sv/stad-och-forvaltning/delta/feedback',
-};
-
-export const DEFAULT_HEADER_MENU_NAME: Record<AppLanguage, string> = {
-  fi:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_FI ??
-    'Sports Helsinki Header FI',
-  en:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_EN ??
-    'Sports Helsinki Header EN',
-  sv:
-    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_SV ??
-    'Sports Helsinki Header SV',
-};
-
-export const DEFAULT_FOOTER_MENU_NAME: Record<AppLanguage, string> = {
-  fi:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_FI ??
-    'Sports Helsinki Footer FI',
-  en:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_EN ??
-    'Sports Helsinki Footer EN',
-  sv:
-    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_SV ??
-    'Sports Helsinki Footer SV',
 };
