--- apps/sports-helsinki/src/domain/footer/footer.module.scss
@@ -1,9 +1,0 @@
-.footer {
-  overflow: hidden;
-  margin-top: calc(var(--spacing-xl) * -1);
-  position: relative;
-  a,
-  a:visited {
-    color: var(--color-white);
-  }
-}
