+++ apps/sports-helsinki/src/pages/search/index.tsx
@@ -1,18 +1,22 @@
-import { NavigationContext, Navigation } from 'events-helsinki-components';
+import {
+  NavigationContext,
+  Navigation,
+  MatomoWrapper,
+  useCommonTranslation,
+} from 'events-helsinki-components';
+import FooterSection from 'events-helsinki-components/components/footer/Footer';
 import type { GetStaticPropsContext } from 'next';
 import React, { useContext } from 'react';
 import { Page as HCRCApolloPage } from 'react-helsinki-headless-cms/apollo';
 import { ROUTES } from '../../constants';
 import getSportsStaticProps from '../../domain/app/getSportsStaticProps';
 import routerHelper from '../../domain/app/routerHelper';
-import FooterSection from '../../domain/footer/Footer';
 import serverSideTranslationsWithCommon from '../../domain/i18n/serverSideTranslationsWithCommon';
-import MatomoWrapper from '../../domain/matomoWrapper/MatomoWrapper';
 import CombinedSearchPage from '../../domain/search/combinedSearch/CombinedSearchPage';
 
 export default function Search() {
   const { footerMenu } = useContext(NavigationContext);
-
+  const { t } = useCommonTranslation();
   return (
     <MatomoWrapper>
       <HCRCApolloPage
@@ -20,7 +24,9 @@ export default function Search() {
         className="pageLayout"
         navigation={<Navigation />}
         content={<CombinedSearchPage defaultTab="Venue" />}
-        footer={<FooterSection menu={footerMenu} />}
+        footer={
+          <FooterSection menu={footerMenu} appName={t('appSports:appName')} />
+        }
       />
     </MatomoWrapper>
   );
