+++ packages/components/src/components/footer/footer.module.scss
