+++ packages/components/src/components/index.ts
@@ -35,8 +35,10 @@ export { default as ToggleButton } from './toggleButton/ToggleButton';
 export { default as TopProgressBar } from './topProgressBar/TopProgressBar';
 export { default as Visible } from './visible/Visible';
 export { default as EllipsedTextWithToggle } from './ellipsedTextWithToggle/EllipsedTextWithToggle';
+export { default as Footer } from './footer/Footer';
 export { default as InfoBlock } from './infoBlock/InfoBlock';
 export { default as MapBox } from './mapBox/MapBox';
+export { default as MatomoWrapper } from './matomoWrapper/MatomoWrapper';
 // export { default as MapView } from './mapView/MapView';
 export { default as Document } from './document/Document';
 export { default as AgeFilter } from './filters/AgeFilter';
