+++ packages/components/src/constants/constants.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import type { AppLanguage } from '../types';
 
 export const MAIN_CONTENT_ID = 'main-content';
 
@@ -58,4 +59,28 @@ export const FEEDBACK_LINKS = {
   fi: 'https://www.hel.fi/helsinki/fi/kaupunki-ja-hallinto/osallistu-ja-vaikuta/palaute',
   en: 'https://www.hel.fi/helsinki/en/administration/participate/feedback',
   sv: 'https://www.hel.fi/helsinki/sv/stad-och-forvaltning/delta/feedback',
+};
+
+export const DEFAULT_HEADER_MENU_NAME: Record<AppLanguage, string> = {
+  fi:
+    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_FI ??
+    'Events Helsinki Header FI',
+  en:
+    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_EN ??
+    'Events Helsinki Header EN',
+  sv:
+    process.env.NEXT_PUBLIC_CMS_HEADER_MENU_NAME_SV ??
+    'Events Helsinki Header SV',
+};
+
+export const DEFAULT_FOOTER_MENU_NAME: Record<AppLanguage, string> = {
+  fi:
+    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_FI ??
+    'Events Helsinki Footer FI',
+  en:
+    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_EN ??
+    'Events Helsinki Footer EN',
+  sv:
+    process.env.NEXT_PUBLIC_CMS_FOOTER_MENU_NAME_SV ??
+    'Events Helsinki Footer SV',
 };
